Move amerge and related functionality to abstract/amerge.py
